+++ lib/state/layout-state.ts
@@ -1,4 +1,4 @@
-import { unstable_ssrSafe as ssrSafe } from "zustand/middleware"
+import { persist, unstable_ssrSafe as ssrSafe } from "zustand/middleware"
 import { immer } from "zustand/middleware/immer"
 import { create } from "zustand"
 
@@ -12,16 +12,19 @@ export interface LayoutState {
 
 export const useLayoutState = create<LayoutState>()(
     ssrSafe(
-        immer((set) => ({
-            showValidationDrawer: false,
-            setShowValidationDrawer(show: boolean) {
-                set({ showValidationDrawer: show })
-            },
+        persist(
+            immer((set) => ({
+                showValidationDrawer: false,
+                setShowValidationDrawer(show: boolean) {
+                    set({ showValidationDrawer: show })
+                },
 
-            showAIAssistant: false,
-            setShowAIAssistant(show: boolean) {
-                set({ showAIAssistant: show })
-            }
-        }))
+                showAIAssistant: true,
+                setShowAIAssistant(show: boolean) {
+                    set({ showAIAssistant: show })
+                }
+            })),
+            { name: "layout-state" }
+        )
     )
 )